+++ discord/objects/__init__.py
@@ -49,6 +49,7 @@ from .poll_answer import PollAnswer
 from .poll_answer_count import PollAnswerCount
 from .poll_media import PollMedia
 from .poll_results import PollResults
+from .resolved_data import ResolvedData
 from .role import Role
 from .role_colors import RoleColors
 from .role_tags import RoleTags