+++ docs/source/changelog.rst
@@ -26,6 +26,7 @@ v26.1.12-dev
 - Added :class:`~discord.enums.PollLayoutType` enum.
 - Added :class:`~discord.objects.PollAnswer` object.
 - Added :class:`~discord.objects.PollMedia` object.
+- Added :class:`~discord.objects.ResolvedData` object.
 
 
 v26.1.11